requirements.txt
Django==4.0.3
djangorestframework==3.13.1
pillow==9.0.1
gunicorn==20.1.0
celery==5.2.3
mysqlclient==2.0.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
django-mysql==4.14.0
redis==4.1.4
opencv-python==4.10.0.84
ultralytics==8.0.124
timm==0.8.13.dev0
yt_dlp
lapx>=0.5.2
opencv-contrib-python==4.10.0.84
diffusers==0.26.3
torch==2.0.0
torchvision==0.15.1
transformers==4.37.1
accelerate==0.26.0
safetensors==0.4.5
einops==0.8.0
onnxruntime-gpu==1.19.2
spaces==0.19.4
omegaconf==2.3.0
peft==0.13.2
huggingface-hub==0.23.2
insightface==0.7.3
gradio==4.44.1
controlnet_aux==0.0.9
gdown==5.2.0
peft
